refactor : TSK-17 SENS 관련 코드 제거
